+++ Runtime/SceneDependencies.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading;
-using AeLa.Utilities.Pool;
+using AeLa.Utilities.Pooling;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
@@ -14,16 +14,79 @@ namespace AeLa.Utilities.SceneDeps
 {
 	public static class SceneDependencies
 	{
+		/// <summary>
+		/// Holds references to dependencies loaded by <see cref="LoadDependenciesAsync"/>.
+		/// Release to decrement the reference count for associated dependencies.
+		/// Dependencies will be unloaded if their reference count is 0 after release.
+		/// </summary>
+		public class Handle : IDisposable
+		{
+			private PooledArray<string> scenes;
+			public Span<string> Scenes => scenes.Span;
+
+			private bool isDisposed;
+
+			public bool IsReleased { get; private set; }
+
+			public bool IsValid => !isDisposed && scenes.Array != null;
+
+			public Handle(PooledArray<string> scenes)
+			{
+				this.scenes = scenes;
+				isDisposed = false;
+				IsReleased = false;
+			}
+
+			/// <summary>
+			/// Releases the handle and unloads any dependency scenes that have no more references.
+			/// </summary>
+			/// <param name="handle"></param>
+			public void Release()
+			{
+				if (IsReleased) return;
+				ReleaseAsync().Forget();
+			}
+
+			/// <summary>
+			/// Releases the handle and unloads any dependency scenes that have no more references.
+			/// </summary>
+			/// <param name="handle"></param>
+			public async UniTask ReleaseAsync()
+			{
+				if (IsReleased || !IsValid) return;
+				IsReleased = true;
+				await ReleaseInternalAsync(this);
+			}
+
+			/// <summary>
+			/// Disposes the handle and releases it if not already released.
+			/// </summary>
+			/// <param name="handle"></param>
+			public void Dispose()
+			{
+				if (isDisposed) return;
+
+				Release();
+				isDisposed = true;
+				scenes.Dispose();
+			}
+		}
+
 		private static Dictionary<string, SceneInstance> pathToInstance = new();
-		private static HashSet<string> loadedDependencies = new();
-		private static HashSet<string> currentDependencies = new();
+
+		/// <summary>
+		/// Loaded dependencies and their reference counts
+		/// </summary>
+		private static Dictionary<string, int> loadedDependencies = new();
+
+		private static Queue<Handle> releaseQueue = new();
 
 		[RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
 		private static void ResetStatic()
 		{
 			pathToInstance = new();
 			loadedDependencies = new();
-			currentDependencies = new();
+			releaseQueue = new();
 		}
 
 		/// <summary>
@@ -31,17 +94,15 @@ namespace AeLa.Utilities.SceneDeps
 		/// in parallel, then activates them once all of their dependencies are loaded and activated.
 		/// </summary>
 		/// <param name="groups">The <see cref="GroupedDependencyChain"/> to load.</param>
-		/// <param name="unloadUnusedDependencies">Whether to unload already loaded dependency scenes that are no longer in use after loading the current groups. If false, you will need to call <see cref="UnloadUnusedDependenciesAsync"/> manually.</param>
-		public static async UniTask LoadDependenciesAsync(
+		/// <returns>A handle that must be used to release references to the dependency scenes</returns>
+		public static async UniTask<Handle> LoadDependenciesAsync(
 			GroupedDependencyChain groups,
-			bool unloadUnusedDependencies = true,
 			CancellationToken ct = default
 		)
 		{
-			currentDependencies.Clear();
-
 			// Below we load and activate the groups in order.
 			// Scenes within a group are loaded/activated simultaneously.
+			using var refs = PooledList<string>.Get();
 			using (ListPool<Exception>.Get(out var exceptions))
 			{
 				for (int i = 0; i < groups.Count; i++)
@@ -54,29 +115,31 @@ namespace AeLa.Utilities.SceneDeps
 					}
 
 					await UniTask.WhenAll(loadScenesTasks);
+
+					refs.List.AddRange(scenes);
 				}
 
 				if (exceptions.Count > 0)
 				{
 					throw new AggregateException(exceptions);
 				}
-
-				if (unloadUnusedDependencies)
-				{
-					await UnloadUnusedDependenciesAsync(ct);
-				}
 			}
 
 			ct.ThrowIfCancellationRequested();
 
-			return;
+			// increment reference counts
+			foreach (var scene in refs)
+			{
+				loadedDependencies[scene]++;
+			}
+
+			return new(refs.ToPooledArray());
 
 			async UniTask LoadDependencyScene(string scene, List<Exception> exceptions)
 			{
 				// already loaded
-				if (loadedDependencies.Contains(scene))
+				if (loadedDependencies.ContainsKey(scene))
 				{
-					currentDependencies.Add(scene);
 					return;
 				}
 
@@ -109,8 +172,7 @@ namespace AeLa.Utilities.SceneDeps
 				}
 
 				var path = sceneInstance.Scene.path;
-				loadedDependencies.Add(path);
-				currentDependencies.Add(path);
+				loadedDependencies.Add(path, 0);
 				pathToInstance[path] = sceneInstance;
 			}
 		}
@@ -120,47 +182,79 @@ namespace AeLa.Utilities.SceneDeps
 		/// Dependencies are determined by calling <see cref="GetDependenciesAsync"/>.
 		/// </summary>
 		/// <param name="scenePath">The path of the scene--this should be the same as the addressables key.</param>
-		/// <param name="unloadUnusedDependencies">Whether to unload already loaded dependency scenes that are no longer in use after loading the current groups. If false, you will need to call <see cref="UnloadUnusedDependenciesAsync"/> manually.</param>
-		public static async UniTask LoadDependenciesAsync(
+		/// <returns>A handle that must be used to release references to the dependency scenes</returns>
+		public static async UniTask<Handle> LoadDependenciesAsync(
 			string scenePath,
-			bool unloadUnusedDependencies = true,
 			CancellationToken ct = default
-		) => await LoadDependenciesAsync(await GetDependenciesAsync(scenePath), unloadUnusedDependencies, ct);
+		) => await LoadDependenciesAsync(await GetDependenciesAsync(scenePath), ct);
 
 		/// <summary>
-		/// Unloads all currently unused dependencies.
+		/// Releases the handle and unloads any dependency scenes that have no more references.
 		/// </summary>
-		/// <param name="ct"></param>
-		public static async UniTask UnloadUnusedDependenciesAsync(CancellationToken ct = default)
+		/// <param name="handle"></param>
+		/// <exception cref="InvalidOperationException">Thrown if scene is not loaded or already has 0 references.</exception>
+		public static UniTask ReleaseAsync(Handle handle)
 		{
-			using var _ = HashSetPool<string>.Get(out var toUnload);
-			toUnload.UnionWith(loadedDependencies);
-			toUnload.ExceptWith(currentDependencies);
+			return handle.ReleaseAsync();
+		}
 
-			using var __ = ListPool<UniTask>.Get(out var tasks);
-			foreach (var scene in toUnload)
-			{
-				tasks.Add(UnloadDependencyAsync(pathToInstance[scene]));
-			}
+		private static UniTask ReleaseInternalAsync(Handle handle)
+		{
+			if (!handle.IsValid) return UniTask.CompletedTask;
 
-			await UniTask.WhenAll(tasks);
+			releaseQueue.Enqueue(handle);
+
+			return releaseQueue.Count == 1 ? ProcessReleaseQueueAsync() : WaitForReleaseQueueAsync();
 		}
 
 		/// <summary>
-		/// Unloads all scene dependencies.
+		/// Waits for any queued handle release operations to finish.
 		/// </summary>
-		public static async UniTask UnloadAllAsync()
+		public static async UniTask WaitForReleaseQueueAsync()
 		{
-			using var _ = ListPool<UniTask>.Get(out var tasks);
-			using var __ = ListPool<string>.Get(out var deps);
-			deps.AddRange(loadedDependencies); // cache loadedDependencies b/c it will be modified by Unload
-			foreach (var scene in deps)
+			while (releaseQueue.Count > 0)
 			{
-				tasks.Add(UnloadDependencyAsync(pathToInstance[scene]));
+				await UniTask.Yield();
 			}
+		}
 
-			await UniTask.WhenAll(tasks);
-			currentDependencies.Clear();
+		private static async UniTask ProcessReleaseQueueAsync()
+		{
+			while (releaseQueue.TryPeek(out var handle))
+			{
+				// update reference counts
+				for (var i = 0; i < handle.Scenes.Length; i++)
+				{
+					var scene = handle.Scenes[i];
+					if (!loadedDependencies.TryGetValue(scene, out var referenceCount))
+					{
+						throw new InvalidOperationException("Attempt to unload dependency scene that is not loaded.");
+					}
+
+					if (referenceCount == 0)
+					{
+						throw new InvalidOperationException("Reference count is already 0");
+					}
+
+					loadedDependencies[scene]--;
+				}
+
+				// unload scenes with no references
+				using var toUnload = PooledList<string>.Get();
+				foreach (var (scene, referenceCount) in loadedDependencies)
+				{
+					if (referenceCount > 0) continue;
+					toUnload.Add(scene);
+				}
+
+				foreach (var scene in toUnload)
+				{
+					await UnloadDependencyAsync(pathToInstance[scene]);
+				}
+
+				handle.Dispose();
+				releaseQueue.Dequeue();
+			}
 		}
 
 		/// <summary>