+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-02-11
+
+### Changed
+
+- BREAKING! `SceneDependencies` now returns a `Handle` object which must be used to release dependency scene references and unload dependency scenes.
+  - Unused dependency scenes are no longer automatically unloaded when `LoadDependenciesAsync` is called
+
+### Removed
+
+- `STM_SceneDependencyManager.UnloadUnusedDependencies`
+
 ## [1.2.0] - 2026-02-11
 
 ### Changed
@@ -12,7 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BREAKING! `SceneDependencyList` is now an abstract base class for `RegexSceneDependencyList`
     - Use Inspector debug mode to update existing asset lists to use `RegexSceneDependencyList.cs`
 - Added `ExplicitSceneDependencyList`
-
 
 ## [1.1.0] - 2026-02-10